Feature/sgc 72

+++ database/migrations/2021_12_29_110000_create_status_table.php
@@ -18,7 +18,7 @@ class CreateStatusTable extends Migration
             $table->string('code')->unique();
             $table->string('name');
             $table->string('description');
-            $table->integer('order');
+            $table->float('order');
             $table->timestamps();
         });
     }
